chore: moving deployment config to inputs dir
